feat: Add search to pipeline nextStepText
